separate methods into diffrent classes
