+++ lib/players.rb
@@ -1,0 +1,10 @@
+class Players
+  def initialize
+    @player1 = 'X'
+    @player2 = 'O'
+  end
+
+  attr_reader :player1
+
+  attr_reader :player2
+end
